Move logging and key vault configuration helpers to out of startup to related projects
