+++ src/Infrastructure/LeanCode.Logging/IHostBuilderExtensions.cs
@@ -1,9 +1,5 @@
 using System.Globalization;
 using System.Reflection;
-using Azure.Core;
-using Azure.Extensions.AspNetCore.Configuration.Secrets;
-using LeanCode.AzureIdentity;
-using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
@@ -11,56 +7,21 @@ using Serilog;
 using Serilog.Events;
 using Serilog.Formatting.Compact;
 
-namespace LeanCode.Components.Startup;
+namespace LeanCode.Logging;
 
 public static class IHostBuilderExtensions
 {
-    public const string SystemLoggersEntryName = LeanProgram.SystemLoggersEntryName;
-
-    public const string VaultUrlKey = "KeyVault:VaultUrl";
+    public const string SystemLoggersEntryName = "Serilog:SystemLoggers";
     public const string MinimumLogLevelKey = "Logging:MinimumLevel";
     public const string EnableDetailedInternalLogsKey = "Logging:EnableDetailedInternalLogs";
     public const string SeqEndpointKey = "Logging:SeqEndpoint";
 
     public const LogEventLevel InternalDefaultLogLevel = LogEventLevel.Warning;
 
-    public static IHostBuilder AddAppConfigurationFromAzureKeyVault(
-        this IHostBuilder builder,
-        TokenCredential? credential = null,
-        string? keyVaultKeyOverride = null,
-        KeyVaultSecretManager? manager = null
-    )
-    {
-        return builder.ConfigureAppConfiguration(
-            (context, builder) =>
-            {
-                ConfigureAzureKeyVault(builder, credential, keyVaultKeyOverride, manager);
-            }
-        );
-    }
-
-    public static IHostBuilder AddAppConfigurationFromAzureKeyVaultOnNonDevelopmentEnvironment(
-        this IHostBuilder builder,
-        TokenCredential? credential = null,
-        string? keyVaultKeyOverride = null,
-        KeyVaultSecretManager? manager = null
-    )
-    {
-        return builder.ConfigureAppConfiguration(
-            (context, builder) =>
-            {
-                if (!context.HostingEnvironment.IsDevelopment())
-                {
-                    ConfigureAzureKeyVault(builder, credential, keyVaultKeyOverride, manager);
-                }
-            }
-        );
-    }
-
     public static IHostBuilder ConfigureDefaultLogging(
         this IHostBuilder builder,
         string projectName,
-        TypesCatalog? destructurers = null,
+        Assembly[]? destructurers = null,
         Action<HostBuilderContext, LoggerConfiguration>? additionalLoggingConfiguration = null
     )
     {
@@ -77,7 +38,7 @@ public static class IHostBuilderExtensions
         this IHostBuilder builder,
         string projectName,
         string appName,
-        TypesCatalog? destructurers = null,
+        Assembly[]? destructurers = null,
         Action<HostBuilderContext, LoggerConfiguration>? additionalLoggingConfiguration = null
     )
     {
@@ -93,7 +54,7 @@ public static class IHostBuilderExtensions
                     .Enrich.WithProperty("project", projectName)
                     .Enrich.WithProperty("app_name", appName)
                     .MinimumLevel.Is(minLogLevel)
-                    .DestructureCommonObjects(destructurers?.Assemblies);
+                    .DestructureCommonObjects(destructurers);
 
                 if (!configuration.GetValue<bool>(EnableDetailedInternalLogsKey))
                 {
@@ -126,34 +87,5 @@ public static class IHostBuilderExtensions
                 logging.AddSerilog();
             }
         );
-    }
-
-    private static void ConfigureAzureKeyVault(
-        IConfigurationBuilder builder,
-        TokenCredential? credential,
-        string? keyVaultUrlKeyOverride,
-        KeyVaultSecretManager? manager
-    )
-    {
-        var configuration = builder.Build();
-
-        var vault = configuration.GetValue<string?>(keyVaultUrlKeyOverride ?? VaultUrlKey);
-        if (vault != null)
-        {
-            var vaultUrl = new Uri(vault);
-            credential ??= DefaultLeanCodeCredential.Create(configuration);
-            if (manager is not null)
-            {
-                builder.AddAzureKeyVault(vaultUrl, credential, manager);
-            }
-            else
-            {
-                builder.AddAzureKeyVault(vaultUrl, credential);
-            }
-        }
-        else
-        {
-            throw new ArgumentException("Application startup exception: null key vault address.");
-        }
     }
 }
